[#38] refactor: Change calendar location to organisms & Chnage initial route of navigator
