+++ src/components/index.ts
@@ -8,7 +8,7 @@ export { default as Modal } from "./modal/modal";
 // export { default as SubmitButton } from "./molecules/auth/submitButton";
 // export { default as CheckTermsButton } from "./molecules/auth/checkTermsButton";
 // export { default as LinkToTerms } from "./link-to-terms/link-to-terms";
-export { default as Calendar } from "./calendar/calendar";
+export { default as Calendar } from "./organisms/calendar/calendar";
 export { default as RecordList } from "./record-list/record-list";
 export { default as Header } from "./header/header";
 export { default as MainHeader } from "./header/mainHeader";
