+++ src/pages/navigator/index.tsx
@@ -61,9 +61,9 @@ export default function Navigator(): ReactElement {
   return (
     <NavigationContainer theme={initialTheme}>
       {(userIn && nickNameIn) || loggedIn ? (
-        <Stack.Navigator initialRouteName="Record" screenOptions={{ headerShown: false }}>
-          <Stack.Screen name="Record" component={Record} />
+        <Stack.Navigator initialRouteName="Index" screenOptions={{ headerShown: false }}>
           <Stack.Screen name="Index" component={DrawerNavigator} />
+          <Stack.Screen name="Record" component={Record} />
           <Stack.Screen name="MonthlyRecord" component={MonthlyRecord} />
           <Stack.Screen name="Profile" component={Profile} />
         </Stack.Navigator>
